% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{add_pxe_defaults}
\alias{add_pxe_defaults}
\title{Add default values to pxe}
\usage{
add_pxe_defaults(pxe)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object}
}
\description{
Add default values to pxe
}
\keyword{internal}
