% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct_functions.R
\docType{class}
\name{api_catalogue}
\alias{api_catalogue}
\alias{update_pxweb_apis}
\alias{api_parameters}
\alias{ApiData}
\alias{MakeUrl}
\alias{base_url}
\alias{get_pxweb_data}
\alias{get_pxweb_dims}
\alias{get_pxweb_levels}
\alias{get_pxweb_metadata}
\alias{pxweb_api-class}
\alias{pxweb_api}
\alias{checkForLevels}
\title{Defunct functions}
\usage{
api_catalogue()

update_pxweb_apis()

api_parameters(url = NULL)

ApiData(
  urlToData,
  ...,
  getDataByGET = FALSE,
  returnMetaData = FALSE,
  returnMetaValues = FALSE,
  returnMetaFrames = FALSE,
  returnApiQuery = FALSE,
  defaultJSONquery = c(1, -2, -1),
  verbosePrint = FALSE,
  use_factors = FALSE,
  urlType = "SSB"
)

MakeUrl(id, urlType = "SSB", getDataByGET = FALSE)

base_url(api, version = NULL, language = NULL)

get_pxweb_data(url, dims, clean = FALSE, encoding = NULL)

get_pxweb_dims(node, verbose = TRUE)

get_pxweb_levels(baseURL, descriptions = FALSE, quiet = FALSE, ...)

get_pxweb_metadata(
  path = NULL,
  node = NULL,
  topnodes = NULL,
  quiet = TRUE,
  baseURL = NULL,
  ...
)

checkForLevels(url)
}
\description{
These function has as from version 0.10.0 become defunct.
Call the functions to get information on new functions to use.
}
\section{Methods}{

\describe{
\item{\code{initialize(get_api = NULL, ...)}}{Create a new pxweb_api object.}
}}

