% Generated by roxygen2 (4.0.1): do not edit by hand
\name{api_timer}
\alias{api_timer}
\title{Creates a timer that keeps track of how many calls that has been called during
a time period depending on the api configuration. If there can't be another call
the function pauses R to do the call.}
\usage{
api_timer(api_url, calls = 1)
}
\arguments{
\item{api_url}{The url to be passed to \link{api_parameters} to get api_configs.}

\item{calls}{The number of calls that the functions should count per function call. Default is one.}
}
\description{
Use this before function that uses api calls.
}
\examples{
url <- base_url("sweSCB", "v1", "sv")
api_timer(url)
}

