% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_database_list.R
\name{as_pxweb_levels}
\alias{as_pxweb_levels}
\title{Assert that x is a correct \code{pxweb_database_list} object.}
\usage{
as_pxweb_levels(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_database_list} object.
}
\keyword{internal}
