% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_api_catalogue.R
\name{assert_pxweb_api_catalogue}
\alias{assert_pxweb_api_catalogue}
\title{Assert a \code{pxweb_api_catalogue} object}
\usage{
assert_pxweb_api_catalogue(x)
}
\arguments{
\item{x}{an object to assert is a \code{pxweb_api_catalogue_entry}.}
}
\description{
Assert a \code{pxweb_api_catalogue} object
}
\keyword{internal}
