% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_assert.R
\name{assert_pxweb_config}
\alias{assert_pxweb_config}
\title{Assert that the config slot is correct}
\usage{
assert_pxweb_config(x)
}
\arguments{
\item{x}{a object to assert}
}
\description{
Assert that the \code{config} slot in the \code{pxweb} object is correct.
}
\keyword{internal}
