% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data.R
\name{assert_pxweb_data}
\alias{assert_pxweb_data}
\title{Assert that x is a correct \code{pxweb_data} object.}
\usage{
assert_pxweb_data(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_data} object.
}
\keyword{internal}
