% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_comments.R
\name{assert_pxweb_data_comments}
\alias{assert_pxweb_data_comments}
\title{Assert that x is a correct \code{pxweb_data_comments} object.}
\usage{
assert_pxweb_data_comments(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_data_comments} object.
}
\keyword{internal}
