% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_jsonstat.R
\name{assert_pxweb_data_jsonstat2}
\alias{assert_pxweb_data_jsonstat2}
\title{Assert that x is a PXWEB API v2 JSON-stat2 data response.}
\usage{
assert_pxweb_data_jsonstat2(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a PXWEB API v2 JSON-stat2 data response.
}
\keyword{internal}
