% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_database_list.R
\name{assert_pxweb_database_list}
\alias{assert_pxweb_database_list}
\title{Assert that x is a correct \code{pxweb_database_list} object.}
\usage{
assert_pxweb_database_list(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_database_list} object.
}
\keyword{internal}
