% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{assert_pxweb_input_allowed}
\alias{assert_pxweb_input_allowed}
\alias{print.pxweb_input_allowed}
\title{Assert a \code{pxweb_input_allowed} object}
\usage{
assert_pxweb_input_allowed(x)

\method{print}{pxweb_input_allowed}(x, ...)
}
\arguments{
\item{x}{an object to assert.}
}
\description{
Assert a \code{pxweb_input_allowed} object
}
\keyword{internal}
