% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_levels.R
\name{assert_pxweb_levels}
\alias{assert_pxweb_levels}
\title{Assert that x is a correct \code{pxweb_levels} object.}
\usage{
assert_pxweb_levels(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_levels} object.
}
\keyword{internal}
