% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{assert_pxweb_metadata}
\alias{assert_pxweb_metadata}
\title{Assert that x is a correct \code{pxweb_metadata} object.}
\usage{
assert_pxweb_metadata(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_metadata} object.
}
\keyword{internal}
