% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{assert_pxweb_metadata_response_v2}
\alias{assert_pxweb_metadata_response_v2}
\title{Assert that x is a PXWEB API v2 metadata response.}
\usage{
assert_pxweb_metadata_response_v2(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a PXWEB API v2 metadata response.
}
\keyword{internal}
