% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{assert_pxweb_query}
\alias{assert_pxweb_query}
\title{Assert a pxweb_query object}
\usage{
assert_pxweb_query(x)
}
\arguments{
\item{x}{an object to assert conferms to the structure of an pxweb_query object.}
}
\description{
Assert a pxweb_query object
}
\keyword{internal}
