% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{assert_pxweb_query_v2}
\alias{assert_pxweb_query_v2}
\title{Assert a \code{pxweb_query_v2} object.}
\usage{
assert_pxweb_query_v2(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert a \code{pxweb_query_v2} object.
}
\keyword{internal}
