% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_assert.R
\name{assert_pxweb_rda_file_path}
\alias{assert_pxweb_rda_file_path}
\title{Assert that the rda_file_path is correct}
\usage{
assert_pxweb_rda_file_path(x)
}
\arguments{
\item{x}{a object to assert}
}
\description{
Assert that the \code{rda_file_path} slot in the \code{pxweb} object is correct.
}
\keyword{internal}
