% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_assert.R
\name{assert_pxweb_url}
\alias{assert_pxweb_url}
\title{Assert that the url structure is correct}
\usage{
assert_pxweb_url(x)
}
\arguments{
\item{x}{a object to assert}
}
\description{
Assert that the \code{url} slot in the \code{pxweb} object is correct.
}
\keyword{internal}
