% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{assert_query_can_be_split_to_batches}
\alias{assert_query_can_be_split_to_batches}
\title{Assert that a given pxweb query can be split}
\usage{
assert_query_can_be_split_to_batches(pxq, pxmd, mxv)
}
\arguments{
\item{pxq}{a [pxweb_query] object}

\item{pxmd}{a [pxweb_metadata] object}

\item{mxv}{maximum batch size}
}
\description{
Assert that a given pxweb query can be split
}
