% Generated by roxygen2 (4.0.1): do not edit by hand
\name{baseURL}
\alias{baseURL}
\title{Return base URL to SCB API}
\usage{
baseURL(version = "v1", lang = "fi", ...)
}
\arguments{
\item{version}{The version of SCB API to use. (Default: \code{v1})}

\item{lang}{The language (two letters) to use in the fetched data. (Default: \code{sv})}

\item{...}{Additional parameters. These are currently ignored.}
}
\description{
...
}
\examples{
a <- baseURL()
b <- baseURL("v2","en")
print(c(a,b))
}

