% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_build_pxweb_config_url.R
\name{build_pxweb_config_url}
\alias{build_pxweb_config_url}
\alias{build_pxweb_config_url.list}
\alias{build_pxweb_config_url.pxweb}
\alias{build_pxweb_config_url.url}
\title{Build the url to get the config from a PXWEB api}
\usage{
build_pxweb_config_url(x)

\method{build_pxweb_config_url}{list}(x)

\method{build_pxweb_config_url}{pxweb}(x)

\method{build_pxweb_config_url}{url}(x)
}
\description{
Build the url to get the config from a PXWEB api
}
\keyword{internal}
