% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_build_pxweb_rda_file_path.R
\name{build_pxweb_rda_file_path}
\alias{build_pxweb_rda_file_path}
\alias{build_pxweb_rda_file_path.character}
\alias{build_pxweb_rda_file_path.url}
\alias{build_pxweb_rda_file_path.pxweb}
\alias{build_pxweb_rda_file_path.list}
\title{Build or get the tmp_rda_file_path from an url or pxweb_api_s3 object}
\usage{
build_pxweb_rda_file_path(x)

\method{build_pxweb_rda_file_path}{character}(x)

\method{build_pxweb_rda_file_path}{url}(x)

\method{build_pxweb_rda_file_path}{pxweb}(x)

\method{build_pxweb_rda_file_path}{list}(x)
}
\arguments{
\item{x}{object to create tmp_file_path for.}
}
\description{
Build or get the tmp_rda_file_path from an url or pxweb_api_s3 object
}
\details{
The hostname is used to
}
\keyword{internal}
