% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_build_pxweb_urls.R
\name{build_pxweb_url}
\alias{build_pxweb_url}
\alias{build_pxweb_url.list}
\alias{build_pxweb_url.pxweb}
\title{Build the url to a PXWEB api}
\usage{
build_pxweb_url(x)

\method{build_pxweb_url}{list}(x)

\method{build_pxweb_url}{pxweb}(x)
}
\description{
Build the url to a PXWEB api
}
\keyword{internal}
