% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_build_pxweb_urls.R
\name{build_pxweb_v2_api_subpath}
\alias{build_pxweb_v2_api_subpath}
\alias{build_pxweb_v2_api_subpath.url}
\title{Build the url to a PXWEB V2 API SUBPATH}
\usage{
build_pxweb_v2_api_subpath(x)

\method{build_pxweb_v2_api_subpath}{url}(x)
}
\description{
Build the url to a PXWEB V2 API SUBPATH
}
\keyword{internal}
