% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildPath}
\alias{buildPath}
\title{Build a path from character elements}
\usage{
buildPath(varname, topnodes = NULL, baseURL, ...)
}
\arguments{
\item{varname}{A character string or a list of strings of the variable(s) in the web API. This can be a data node or a tree node.}

\item{topnodes}{A string or a list of strings containing the top nodes \emph{in top-to-bottom order}}

\item{baseURL}{The base URL to use, depending on the web service.}

\item{...}{Further arguments passed to  \code{base_url()}.}
}
\description{
This function takes a list of strings and builds a URL to the PX-web API \emph{in reverser order}.
}

