% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_pxweb_data_internal.R
\name{calc_dim_type}
\alias{calc_dim_type}
\title{Calculates which dimensions that are used as row and col variables in data}
\usage{
calc_dim_type(dim_data2clean, dim_size)
}
\arguments{
\item{dim_data2clean}{The dimension of the data from PXWEB}

\item{dim_size}{for the call from \link{get_dim_size}}
}
\value{
list with index (row and col dim)
}
\description{
Calculates which dimensions that are used as row and col variables in data
}

