% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{calculate_data_dim}
\alias{calculate_data_dim}
\title{The function calculates the dimension of a data object}
\usage{
calculate_data_dim(dim_length, clean)
}
\arguments{
\item{dim_length}{Vector from \link{get_dim_size}}

\item{clean}{Is the data cleaned (TRUE/FALSE)}
}
\value{
An integer vector of length 2
}
\description{
The function calculates the dimension of a data object
}
\examples{
url <- "http://api.scb.se/OV0104/v1/doris/sv/ssd/BE/BE0101/BE0101A/BefolkningNy"
dims <- list(Region = c('*'), Civilstand = c('*'), Alder = '1', Kon = c('*'), 
            ContentsCode = c('*'), Tid = c('*'))
}
\keyword{internal}
