% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{checkForLevels}
\alias{checkForLevels}
\title{Function to silently test for existence of metadata and return TRUE or FALSE based on the result of that test.}
\usage{
checkForLevels(url)
}
\arguments{
\item{url}{API url}
}
\description{
Function to silently test for existence of metadata and return TRUE or FALSE based on the result of that test.
}

