% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_pxweb_internal.R
\name{choose_pxweb_api}
\alias{choose_pxweb_api}
\title{Choose an api from api_catalogue}
\usage{
choose_pxweb_api()
}
\value{
base url to the specific data base
}
\description{
Choose an api from api_catalogue
}
\keyword{internal}
