% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_pxweb_internal.R
\name{choose_pxweb_database_url}
\alias{choose_pxweb_database_url}
\title{Calculate a specific database to get data from}
\usage{
choose_pxweb_database_url(baseURL, pre_choice = NULL)
}
\arguments{
\item{baseURL}{The basic url to the pxweb api}

\item{pre_choice}{Predifined choice of database}
}
\value{
base url to the specific data base
}
\description{
Calculate a specific database to get data from
}

