% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_pxweb}
\alias{clean_pxweb}
\title{Clean raw data from PX-WEB}
\usage{
clean_pxweb(data2clean, head, url)
}
\arguments{
\item{data2clean}{Data to clean.}

\item{head}{Full variable names as character vector}

\item{url}{url to the bottom nod (to get meta data)}
}
\value{
data frame melted and in R (numeric) format
}
\description{
This function clean the raw data from PX-WEB to R tall format.
}

