% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findData.Download}
\alias{findData.Download}
\title{Traverse node for query alternatives and download data.}
\usage{
findData.Download(dataNode, test_input = NULL, ...)
}
\arguments{
\item{dataNode}{Botton node in SCB node tree.}

\item{test_input}{Vector of length 4 to test inputs to the first 4 questions in the query.}

\item{...}{further parameters. These are currently ignored.}
}
\description{
Goes through the dataNode and ask user for input for all
variables and then put this together to a query for \link{scbGetData}.
}

