% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_pxweb_internal.R
\name{findData.input}
\alias{findData.input}
\title{Get input that is consistent with}
\usage{
findData.input(type, input = NULL, test_input = character(0),
  silent = FALSE)
}
\arguments{
\item{type}{type of input to get.}

\item{input}{data.frame with input data to use with}

\item{test_input}{input for test cases}

\item{silent}{no output}
}
\description{
Get input that is consistent with
}

