% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findData}
\alias{findData}
\title{Find and download data from PX-WEB API}
\usage{
findData(baseURL, history = FALSE, ...)
}
\arguments{
\item{baseURL}{The base URL to use, depending on the web service.}

\item{history}{keep the history when the function is running.}

\item{...}{further parameters. These are currently ignored.}
}
\description{
Wrapper function (for \link{scbGetData} and \link{scbGetMetadata}) to simply find and download data from SCB to the current R session.
}
\examples{
\dontrun{
api_parameters() # List options
baseURL <- base_url("sweSCB", "v1", "sv")
d <- findData(baseURL)
}
}
\seealso{
\code{\link{scbGetMetadata}}, \code{\link{scbGetData}}
}

