% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{generate_permutations}
\alias{generate_permutations}
\title{Generate batch permutations}
\usage{
generate_permutations(x)
}
\arguments{
\item{x}{a vector with elements to permute}
}
\description{
Generate batch permutations
}
\details{
Generates permutations of dim. If more than 6 dim (highly unlikely) a sample of 1000 combinations
is drawn. Otherwise all possible permutations are returned.
}
\keyword{internal}
