% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_catalogue.R
\name{get_api_file_path}
\alias{get_api_file_path}
\title{Refactorized function to get api_file path and copy if needed}
\usage{
get_api_file_path()
}
\description{
Refactorized function to get api_file path and copy if needed
}
\keyword{internal}
