% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_catalogue.R
\name{get_api_list}
\alias{get_api_list}
\title{Get the api catalogue from json file}
\usage{
get_api_list(raw = FALSE)
}
\arguments{
\item{raw}{Get the raw list from the api json file.}
}
\value{
api_list object
}
\description{
Get the api catalogue from json file
}
\keyword{internal}

