% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{get_dim_size}
\alias{get_dim_size}
\title{The function calculates the dimension of a query. That means the numberof
categories for each variable downloaded}
\usage{
get_dim_size(url, dims, content_node = NULL)
}
\arguments{
\item{url}{The url the query uses}

\item{dims}{The dimension object to use for downloading}

\item{content_node}{Used if already a content node has been downloaded}
}
\value{
list with two slots 
[[1]] containes named vector with number of slots
[[2]] containes content_node got with \link{get_pxweb_metadata}
}
\description{
The function calculates the dimension of a query. That means the numberof
categories for each variable downloaded
}
\keyword{internal}
