% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_api_catalogue.R
\name{get_github_api_urls}
\alias{get_github_api_urls}
\title{Get (hard coded) api catalogue github urls}
\usage{
get_github_api_urls(type)
}
\arguments{
\item{type}{which type of github url is needed ('github_api' or 'github_raw').}
}
\description{
Get (hard coded) api catalogue github urls
}
\keyword{internal}
