% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_pxweb_levels.R
\name{get_pxweb_levels}
\alias{get_pxweb_levels}
\title{Get levels from API node}
\usage{
get_pxweb_levels(baseURL, descriptions = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{baseURL}{The base URL to use, depending on the web service.}

\item{descriptions}{Whether to include node descriptions with the list of node IDs. (default: \code{FALSE})}

\item{quiet}{Quiet mode. Whether to stop with an error if the input node does not contain any subnodes. If set to \code{TRUE}, the function will quietly return FALSE without any errors. (default: \code{FALSE})}

\item{...}{further argument to send to \code{get_pxweb_metadata}}
}
\description{
Get levels from a node in the API. If at the lowest node, return a warning.
}

