% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_pxweb_metadata}
\alias{get_pxweb_metadata}
\title{Get data}
\usage{
get_pxweb_metadata(path = NULL, node = NULL, topnodes = NULL,
  quiet = TRUE, baseURL = NULL, ...)
}
\arguments{
\item{path}{URL to fetch metadata from. If left empty, the function constructs a URL from the \code{node} and \code{topnodes} arguments}

\item{node}{A string with the name of the node to fetch. This is ignored if \code{path} is supplied.}

\item{topnodes}{A string or list of strings with the names of the parent nodes of \code{node}. This is ignored if \code{path} is supplied.}

\item{quiet}{Quiet mode (never return a message to the user)}

\item{baseURL}{The base URL to use, depending on the web service. Needed if path argument is not provided.}

\item{...}{Further arguments passed to  \code{base_url()}.}
}
\description{
Get data from the API. If at the lowest node, provide the user with a friendly message about this.
}
\examples{
api_url <- base_url("api.scb.se", language = "sv")
\donttest{
top_node <- get_pxweb_metadata(api_url)

bottom_node <- get_pxweb_metadata("http://api.scb.se/OV0104/v1/doris/sv/ssd/AM/AM0114/LCIArbKv")
}
}

