% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_utils_internal.R
\name{getContent}
\alias{getContent}
\title{Get content from response}
\usage{
getContent(response, type = "csv")
}
\arguments{
\item{response}{response object}

\item{type}{type format}
}
\description{
Get the content from a response object
}
\keyword{internal}
