% Generated by roxygen2 (4.0.1): do not edit by hand
\name{interactive_pxweb}
\alias{interactive_pxweb}
\title{Find and download data interactively from PX-WEB API}
\usage{
interactive_pxweb(baseURL, history = FALSE, ...)
}
\arguments{
\item{baseURL}{The base URL to use, depending on the web service.}

\item{history}{keep the history when the function is running.}

\item{...}{further parameters. These are currently ignored.}
}
\description{
Wrapper function (for \link{get_pxweb_data} and \link{get_pxweb_metadata}) to simply find and download data to the current R session.
}
\examples{
api_parameters() # List options
baseURL <- base_url("sweSCB", "v1", "sv")
\dontrun{
d <- interactive_pxweb(baseURL)
}
}
\seealso{
\code{\link{get_pxweb_metadata}}, \code{\link{get_pxweb_data}}
}

