% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{is_pxweb_query_selection}
\alias{is_pxweb_query_selection}
\title{Test for PXWEB API v2 query selection helpers.}
\usage{
is_pxweb_query_selection(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Test for PXWEB API v2 query selection helpers.
}
\keyword{internal}
