% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_progress_bar.R
\docType{class}
\name{msg_progress_bar-class}
\alias{msg_progress_bar}
\alias{msg_progress_bar-class}
\title{Message progress bar}
\description{
A simple progress bar to use in R packages where messages are prefered to console output.
}
\section{Fields}{

\describe{
\item{\code{iter}}{Total number of iterations}

\item{\code{i}}{Current iteration}

\item{\code{width}}{Width of the R console}

\item{\code{width_bar}}{Width of the progress bar}

\item{\code{progress}}{The number of character printed (continous)}

\item{\code{progress_step}}{Addition to progress per iteration}
}}
\section{Methods}{

\describe{
\item{\code{increment()}}{A messagebar object.}

\item{\code{initialize(iter)}}{Initialize a messagebar object}
}}
\examples{
 test_bar <- function(iter = 10){
 bar <- msg_progress_bar(iter)
 for(j in 1:iter){
   bar$increment()
   Sys.sleep(0.01)
   }
 }
 test_bar(100)
  
}
\author{
Mans Magnusson (MansMeg @ github)
}

