% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_ApiData.R
\name{Number}
\alias{Number}
\title{Adding leading zeros}
\usage{
Number(n, width = 3)
}
\arguments{
\item{n}{numeric vector}

\item{width}{width}
}
\value{
Number as string
}
\description{
Adding leading zeros
}
\keyword{internal}
