% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_parse_url_or_fail.R
\name{parse_url_or_fail}
\alias{parse_url_or_fail}
\title{Parse a character string or throws error if it fails}
\usage{
parse_url_or_fail(x)
}
\arguments{
\item{x}{a character element to parse}
}
\description{
Parse a character string or throws error if it fails
}
\keyword{internal}
