% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{permutations}
\alias{permutations}
\title{Generate permutations of dimensions to find optimal no of batches}
\usage{
permutations(n, r, v = 1:n, set = TRUE, repeats.allowed = FALSE)
}
\arguments{
\item{n}{See \code{gtools::permutations}.}

\item{r}{See \code{gtools::permutations}.}

\item{v}{See \code{gtools::permutations}.}

\item{repeats.allowed}{See \code{gtools::permutations}.}

\item{nset}{See \code{gtools::permutations}.}
}
\description{
Generate permutations of dimensions to find optimal no of batches
}
\details{
Taken from gtools to minimize dependencies. See permutations 
of the gtools packages for details
}
\keyword{internal}
