% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_api_catalogue.R
\name{print.pxweb_api_catalogue_entry}
\alias{print.pxweb_api_catalogue_entry}
\title{Print a catalogue entry}
\usage{
\method{print}{pxweb_api_catalogue_entry}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a catalogue entry
}
