% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data.R
\name{pxd_values_to_valuetexts}
\alias{pxd_values_to_valuetexts}
\title{Convert a pxweb data objects values to valuetext}
\usage{
pxd_values_to_valuetexts(x, variable, value_levels)
}
\arguments{
\item{x}{a \code{pxweb_data} object}

\item{variable}{a variable to convert}

\item{value_levels}{value_levels to convert}
}
\description{
Convert a pxweb data objects values to valuetext
}
\keyword{internal}
