% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_allowed_input}
\alias{pxe_allowed_input}
\title{Defines allowed input for a position in a \code{pxweb_explorer} object.}
\usage{
pxe_allowed_input(pxe)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object to get allowed input for.}
}
\description{
Defines allowed input for a position in a \code{pxweb_explorer} object.
}
\keyword{internal}
