% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_back_position}
\alias{pxe_back_position}
\alias{pxe_add_position}
\title{Move in the \code{pxweb_explorer} position}
\usage{
pxe_back_position(pxe)

pxe_add_position(pxe, new_pos)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object.}

\item{new_pos}{add a new position.}
}
\description{
Move in the \code{pxweb_explorer} position
}
\keyword{internal}
