% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_data_url}
\alias{pxe_data_url}
\title{Get the url to a table}
\usage{
pxe_data_url(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object}
}
\description{
Get the url to a table
}
\keyword{internal}
