% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_handle_input}
\alias{pxe_handle_input}
\alias{pxe_handle_input.numeric}
\alias{pxe_handle_input.character}
\title{Handle a user input for a \code{pxweb_explorer} object.}
\usage{
pxe_handle_input(user_input, pxe)

\method{pxe_handle_input}{numeric}(user_input, pxe)

\method{pxe_handle_input}{character}(user_input, pxe)
}
\arguments{
\item{user_input}{an (allowed) user input to handle.}

\item{pxe}{a \code{pxweb_explorer} object to get user input for.}
}
\description{
Handle a user input for a \code{pxweb_explorer} object.
}
\seealso{
pxe_allowed_input()
}
\keyword{internal}
