% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_input}
\alias{pxe_input}
\alias{pxe_parse_input}
\title{Get (allowed) inputs for a \code{pxweb_input_allowed} object.}
\usage{
pxe_input(allowed_input, title = NULL, test_input = NULL)

pxe_parse_input(user_input, allowed_input)
}
\arguments{
\item{allowed_input}{a \code{pxweb_input_allowed}.}

\item{title}{Print (using cat()) before ask for the allowed choices.}

\item{test_input}{supplying a test input (for testing only)}
}
\description{
Get (allowed) inputs for a \code{pxweb_input_allowed} object.
}
\details{
It handles input and checks if the input is allowed.
}
\keyword{internal}
