% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_interactive_get_data}
\alias{pxe_interactive_get_data}
\title{Ask to download and download data}
\usage{
pxe_interactive_get_data(pxe, test_input = NULL)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object with full query}

\item{test_input}{a test input for testing the function.
Since two question, supply a vector of length two.}
}
\description{
Ask to download and download data
}
\keyword{internal}
