% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_metadata_choices}
\alias{pxe_metadata_choices}
\alias{pxe_metadata_choices<-}
\title{Get and set pxe_metadata_coices}
\usage{
pxe_metadata_choices(x)

pxe_metadata_choices(x) <- value
}
\arguments{
\item{x}{a \code{pxweb_explorer} object}

\item{value}{an object to set as pxe_metadata_choice}
}
\description{
Get and set pxe_metadata_coices
}
\keyword{internal}
