% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_metadata_variable_names}
\alias{pxe_metadata_variable_names}
\title{Get the meta data variable names from a \code{pxweb_explorer} object.}
\usage{
pxe_metadata_variable_names(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object}
}
\description{
Get the meta data variable names from a \code{pxweb_explorer} object.
}
\keyword{internal}
