% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_choice_size}
\alias{pxe_position_choice_size}
\alias{pxe_position_print_size}
\title{How many choices has the current position?}
\usage{
pxe_position_choice_size(x)

pxe_position_print_size(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
How many choices has the current position?
}
\keyword{internal}
