% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_is_api_catalogue}
\alias{pxe_position_is_api_catalogue}
\title{Is the current position an api_catalogue position?}
\usage{
pxe_position_is_api_catalogue(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
Is the current position an api_catalogue position?
}
\keyword{internal}
