% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_is_full_query}
\alias{pxe_position_is_full_query}
\title{Is the current position a full query (i.e. choices for all metadata variables)?}
\usage{
pxe_position_is_full_query(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
Is the current position a full query (i.e. choices for all metadata variables)?
}
\keyword{internal}
