% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_is_metadata}
\alias{pxe_position_is_metadata}
\title{Is the current position a metadata object?}
\usage{
pxe_position_is_metadata(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
Is the current position a metadata object?
}
\keyword{internal}
