% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_multiple_choice_allowed}
\alias{pxe_position_multiple_choice_allowed}
\title{Are multiple choices allowed?}
\usage{
pxe_position_multiple_choice_allowed(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
Are multiple choices allowed?
}
\keyword{internal}
