% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_title}
\alias{pxe_position_title}
\title{Get the table title for the current position}
\usage{
pxe_position_title(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object.}
}
\description{
Get the table title for the current position
}
\keyword{internal}
