% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_position_variable_can_be_eliminated}
\alias{pxe_position_variable_can_be_eliminated}
\title{Can the variable at the current position be eliminated?}
\usage{
pxe_position_variable_can_be_eliminated(x)
}
\arguments{
\item{x}{a \code{pxweb_explorer} object to check.}
}
\description{
Can the variable at the current position be eliminated?
}
\keyword{internal}
