% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_print_download_code}
\alias{pxe_print_download_code}
\title{Print code to download query}
\usage{
pxe_print_download_code(pxe, as)
}
\arguments{
\item{pxe}{a \code{pxweb_query} object.}

\item{as}{\code{json} or \code{r}.}
}
\description{
Print code to download query
}
\keyword{internal}
