% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_pxobj_at_position}
\alias{pxe_pxobj_at_position}
\alias{pxe_pxobj_at_position<-}
\title{Return the pxweb object at the current position}
\usage{
pxe_pxobj_at_position(x)

pxe_pxobj_at_position(x) <- value
}
\arguments{
\item{x}{a \code{pxweb_explorer} object.}
}
\description{
Return the pxweb object at the current position
}
\keyword{internal}
