% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb.R
\name{pxweb}
\alias{pxweb}
\alias{is.pxweb}
\alias{print.pxweb}
\title{S3 constructor for \code{pxweb} api object.}
\usage{
pxweb(url)

is.pxweb(x)

\method{print}{pxweb}(x, ...)
}
\arguments{
\item{url}{an url to a pxweb api including language and version of the api. See examples.}

\item{x}{an an object to test if it is a \code{pxweb} object.}

\item{...}{further arguments supplied to \code{print()}.}
}
\value{
A \code{pxweb} object with slots:
FIX!
}
\description{
The pxwebapi object contain all information to do calls to the pxweb api and keep count 
of the number of calls. The object is constructed
The object will also be cached in R temp folder to minimize calls to api.
All urls should be passed through the constructor to set up the pxweb api config.

\emph{Garantuees}:
The base_url has been pinged
The sub_path has been checked
The config has been captured from the API
The url has been checked to be a pxweb api (through config)
}
\examples{
pxapi_1 <- pxweb(url ="http://api.scb.se/OV0104/v1/doris/sv/ssd/START/ME/ME0104/ME0104C/ME0104T24")
pxapi_2 <- pxweb(url ="http://api.scb.se/OV0104/v1/doris/sv")

}
