% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_add_api_subpath.R
\name{pxweb_add_api_subpath}
\alias{pxweb_add_api_subpath}
\title{Add the subpath slot to a pxweb path slot}
\usage{
pxweb_add_api_subpath(obj)
}
\arguments{
\item{obj}{an object to add subpath to}
}
\description{
Add the subpath slot to a pxweb path slot
}
\details{
Queries the path from pos 1 and up until a config is returned.
}
\keyword{internal}
