% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_add_call.R
\name{pxweb_add_call}
\alias{pxweb_add_call}
\title{Add an api call to a pxweb_api_s3 object}
\usage{
pxweb_add_call(obj, time_stamp = Sys.time())
}
\arguments{
\item{obj}{a \code{pxweb_api_s3} object}
}
\description{
The pxweb_add_call function add a call the the api in the call stack, then compute 

Promise: 
The stored rda api object will always have the latest calls
This is not thread safe so only one session at a time should call the api.
}
\keyword{internal}
