% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_get.R
\name{pxweb_add_mandatory_variables}
\alias{pxweb_add_mandatory_variables}
\title{Add mandatory variables to query}
\usage{
pxweb_add_mandatory_variables(pxq, pxmd)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{pxmd}{a \code{pxweb_metadata} object.}
}
\description{
Add mandatory variables to query
}
\details{
Complement queries that lack explicit requests for variables with requests for every value of these variables.
}
\keyword{internal}
