% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_api_catalogue.R
\name{pxweb_api_catalogue_entry}
\alias{pxweb_api_catalogue_entry}
\alias{pxweb_api_catalogue_entry.list}
\alias{assert_pxweb_api_catalogue_entry}
\title{Constructor for \code{pxweb_api_catalogue_entry}.}
\usage{
pxweb_api_catalogue_entry(x)

\method{pxweb_api_catalogue_entry}{list}(x)

assert_pxweb_api_catalogue_entry(x)
}
\arguments{
\item{x}{an object to convert to a \code{pxweb_api_catalogue_entry} object.}

\item{x}{an object to assert is a \code{pxweb_api_catalogue_entry}.}
}
\description{
Constructor for \code{pxweb_api_catalogue_entry}.
}
\keyword{internal}
