% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_clear_cache.R
\name{pxweb_clear_cache}
\alias{pxweb_clear_cache}
\title{Clear cache of all (or one) \code{pxweb} object}
\usage{
pxweb_clear_cache(x = NULL)
}
\arguments{
\item{x}{a \code{pxweb} object to clear cache for. Default is NULL (clear everything).}
}
\description{
Clear cache of all (or one) \code{pxweb} object
}
\details{
Clean the cache and sleep to restore api timing limit.
}
\keyword{internal}
