% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data.R
\name{pxweb_data}
\alias{pxweb_data}
\title{Construct a \code{pxweb_data} object.}
\usage{
pxweb_data(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API to convert to a \code{pxweb_data} object.}
}
\value{
a \code{pxweb_data} object
}
\description{
An object that contain the data for a given PXWEB table.
}
\keyword{internal}
