% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_c.R
\name{pxweb_data_c}
\alias{pxweb_data_c}
\title{Combine pxweb objects}
\usage{
pxweb_data_c(x)
}
\arguments{
\item{x}{a list with \code{pxweb} objects.}
}
\description{
Combine pxweb objects
}
\keyword{internal}
