% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_comments.R
\name{pxweb_data_column_comment}
\alias{pxweb_data_column_comment}
\alias{pxweb_data_value_comment}
\alias{pxweb_data_obs_comment}
\title{Construct a \code{pxweb_data_comment} object}
\usage{
pxweb_data_column_comment(x, column_idx)

pxweb_data_value_comment(x, comment_idx)

pxweb_data_obs_comment(x, obs_idx)
}
\arguments{
\item{x}{an \code{pxweb_data} to extract a \code{pxweb_data_comment} object from.}

\item{column_idx}{the index of the column to extract.}

\item{comment_idx}{the index of the comment to extract.}

\item{obs_idx}{the index of the comment to extract.}
}
\description{
Construct a \code{pxweb_data_comment} object
}
\keyword{internal}
