% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_jsonstat.R
\name{pxweb_data_jsonstat2}
\alias{pxweb_data_jsonstat2}
\title{Construct a \code{pxweb_data_jsonstat2} object.}
\usage{
pxweb_data_jsonstat2(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API v2 data endpoint.}
}
\value{
a \code{pxweb_data_jsonstat2} object.
}
\description{
An object that contains JSON-stat2 data returned by a PXWEB API v2
\code{/tables/{tableId}/data} endpoint.
}
\keyword{internal}
