% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_comments.R
\name{pxweb_data_v2_column_comment}
\alias{pxweb_data_v2_column_comment}
\title{Construct a v2 dimension note comment.}
\usage{
pxweb_data_v2_column_comment(x, column_idx, comment)
}
\arguments{
\item{x}{a \code{pxweb_data_v2} object.}

\item{column_idx}{integer index of the dimension column.}

\item{comment}{a normalized comment string.}
}
\value{
a \code{pxweb_data_comment} object.
}
\description{
Converts a JSON-stat2 dimension-level note to a v1-style
\code{column_comment} object.
}
\keyword{internal}
