% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_database_list.R
\name{pxweb_database_list}
\alias{pxweb_database_list}
\title{Construct a \code{pxweb_database_list} object.}
\usage{
pxweb_database_list(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API to convert to a \code{pxweb_database_list} object.}
}
\value{
a \code{pxweb_database_list} object
}
\description{
An object that contain the list of databases for a given PXWEB api.
}
\keyword{internal}
