% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxweb_explorer}
\alias{pxweb_explorer}
\alias{pxweb_explorer.NULL}
\alias{pxweb_explorer.character}
\alias{pxweb_explorer.pxweb}
\alias{pxweb_explorer.pxweb_api_catalogue_entry}
\alias{assert_pxweb_explorer}
\title{Create a \code{pxweb_explorer} object.}
\usage{
pxweb_explorer(x = NULL)

\method{pxweb_explorer}{NULL}(x)

\method{pxweb_explorer}{character}(x)

\method{pxweb_explorer}{pxweb}(x)

\method{pxweb_explorer}{pxweb_api_catalogue_entry}(x)

assert_pxweb_explorer(x)
}
\arguments{
\item{x}{a \code{pxweb} object, a PXWEB url, \code{NULL} or an api in the api catalogue.

\dontrun{
 d <- pxweb_explorer()
 d <- pxweb_explorer(x = "api.scb.se")
 d <- pxweb_explorer(x = "http://api.scb.se/OV0104/v1/doris/en/ssd/BE/BE0101/")
}}
}
\description{
Create a \code{pxweb_explorer} object.
}
\keyword{internal}
