% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_test_api_endpoint.R
\name{pxweb_get_api_test_data_frame}
\alias{pxweb_get_api_test_data_frame}
\title{Build api test data.frame}
\usage{
pxweb_get_api_test_data_frame(x)
}
\arguments{
\item{x}{a pxweb object}
}
\description{
Build api test data.frame
}
\keyword{internal}
