% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_logs.R
\name{pxweb_http_log_on}
\alias{pxweb_http_log_on}
\alias{pxweb_http_log_off}
\alias{pxweb_log_paths_path}
\alias{pxweb_http_log_is_on}
\alias{pxweb_log_http_response}
\alias{pxweb_response_to_log_as_json}
\title{Setup a structure to log all API calls}
\usage{
pxweb_http_log_on(fn = "log_pxweb_api_http_calls.txt")

pxweb_http_log_off()

pxweb_log_paths_path()

pxweb_http_log_is_on()

pxweb_log_http_response(r)

pxweb_response_to_log_as_json(r)
}
\arguments{
\item{fn}{The file name of the log file.}
}
\description{
Setup a structure to log all API calls
}
\details{
Set up internal structure to handle http PXWEB API calls. http calls are stored in a log file.
To access the path to this log file, the path is stored in the pxweb folder in the tempdir(). 
The path is given by \code{pxweb_log_paths_path()}.
If the file exists, the http calls should be logged (appended) to the log file where the object contains the path.
}
\keyword{internal}
