% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxweb_interactive}
\alias{pxweb_interactive}
\alias{interactive_pxweb}
\title{Find and download data interactively from a PXWEB API}
\usage{
pxweb_interactive(x = NULL)

interactive_pxweb(x = NULL)
}
\arguments{
\item{x}{The name or alias of the pxweb api to connect to, a \code{pxweb} object or an url.}
}
\value{
Invisibly returns a list with at least:
\describe{
  \item{\code{url}}{The URL to the selected data table.}
  \item{\code{query}}{A \code{pxweb_query} object that can be passed to \code{\link{pxweb_get}}.}
}
If data is downloaded during the interactive session, the list also includes
\code{data}, containing the downloaded data object returned by \code{\link{pxweb_get}}.
This return value is not the same as calling \code{pxweb_get()} directly.
}
\description{
Wrapper function (for \link{pxweb_get}) to simply find and download data to the current R session.
\code{interactive_pxweb()} is a deprecated alias for \code{pxweb_interactive()}.
}
\examples{
pxweb_api_catalogue() # List apis

## The examples below can only be run in interactive mode
##  x <- pxweb_interactive()
##  x <- pxweb_interactive(x = "api.scb.se")
##  x <- pxweb_interactive(x = "https://api.scb.se/OV0104/v1/doris/en/ssd/BE/BE0101/")
##  x <- pxweb_interactive(x = "https://api.scb.se/OV0104/v1/doris/en/ssd/BE/BE0101/BE0101A/")

}
\seealso{
\code{\link{pxweb_get}}
}
