% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxweb_interactive_input}
\alias{pxweb_interactive_input}
\title{Get input from user}
\usage{
pxweb_interactive_input(pxe)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object to get user input for.}
}
\description{
Get input from user
}
\keyword{internal}
