% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_levels.R
\name{pxweb_levels}
\alias{pxweb_levels}
\title{Construct a \code{pxweb_levels} object.}
\usage{
pxweb_levels(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API to convert to a \code{pxweb_levels} object.}
}
\value{
a \code{pxweb_levels} object
}
\description{
An object that contain the levels for a given PXWEB api position.
}
\keyword{internal}
