% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_levels.R
\name{pxweb_levels_remove_headers}
\alias{pxweb_levels_remove_headers}
\title{Remove headers from PXWEB levels object}
\usage{
pxweb_levels_remove_headers(x)
}
\arguments{
\item{x}{a \code{pxweb_levels} object}
}
\description{
Remove headers from PXWEB levels object
}
\keyword{internal}
