% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{pxweb_metadata}
\alias{pxweb_metadata}
\title{Construct a \code{pxweb_metadata} object.}
\usage{
pxweb_metadata(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API to convert to a \code{pxweb_metadata} object.}
}
\value{
a \code{pxweb_metadata} object
}
\description{
An object that contain the metadata for a given PXWEB table.
}
\keyword{internal}
