% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{pxweb_metadata_dim}
\alias{pxweb_metadata_dim}
\title{Compue the dimension of a metadata object}
\usage{
pxweb_metadata_dim(pxmd)
}
\arguments{
\item{pxmd}{a \code{pxweb_metadata} object.}
}
\description{
Compue the dimension of a metadata object
}
\keyword{internal}
