% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{pxweb_metadata_time}
\alias{pxweb_metadata_time}
\title{Get boolean vector}
\usage{
pxweb_metadata_time(pxmd)
}
\arguments{
\item{pxmd}{a \code{pxweb_metadata} object.}
}
\value{
pxweb_metadata eliminations as a named boolean vector.
}
\description{
Get boolean vector
}
\keyword{internal}
