% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{pxweb_metadata_v2_value_texts}
\alias{pxweb_metadata_v2_value_texts}
\title{Get value texts from a PXWEB API v2 dimension.}
\usage{
pxweb_metadata_v2_value_texts(dim, values)
}
\arguments{
\item{dim}{a dimension list from a PXWEB API v2 metadata response.}

\item{values}{a character vector of value codes.}
}
\value{
a character vector of value labels aligned with \code{values}.
}
\description{
Get value texts from a PXWEB API v2 dimension.
}
\keyword{internal}
