% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_metadata.R
\name{pxweb_metadata_v2_values}
\alias{pxweb_metadata_v2_values}
\title{Get ordered values from a PXWEB API v2 dimension.}
\usage{
pxweb_metadata_v2_values(dim)
}
\arguments{
\item{dim}{a dimension list from a PXWEB API v2 metadata response.}
}
\value{
a character vector of value codes ordered by the dimension index.
}
\description{
Get ordered values from a PXWEB API v2 dimension.
}
\keyword{internal}
