% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_parse_response.R
\name{pxweb_parse_response}
\alias{pxweb_parse_response}
\title{Parse the response from a PXWEB API}
\usage{
pxweb_parse_response(x)
}
\arguments{
\item{x}{a \code{httr} response object from a PXWEB call.}
}
\description{
Parse the response from a PXWEB API
}
\keyword{internal}
