% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pxweb_query_class.R
\docType{class}
\name{pxweb_query-class}
\alias{pxweb_query}
\alias{pxweb_query-class}
\title{A Reference Class to represent an pxweb api query}
\description{
A Reference Class to represent an pxweb api query
}
\section{Fields}{

\describe{
\item{\code{query}}{}
}}
\section{Methods}{

\describe{
\item{\code{get_query()}}{Get data from content.}

\item{\code{initialize(query, pxwebapi)}}{Create a new pxwebapi_query}

\item{\code{parse_query(query)}}{Parse pxweb query}

\item{\code{parse_query_json(query)}}{Parse json pxweb query}

\item{\code{parse_query_list(query)}}{Parse list pxweb query}
}}
\examples{
  
  
}
\references{
http://www.scb.se/Grupp/OmSCB/API/API-description.pdf
}

