% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_as_json}
\alias{pxweb_query_as_json}
\title{Convert a \code{pxweb_query} object to a \code{json} string}
\usage{
pxweb_query_as_json(pxq, ...)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{...}{further argument to \code{jsonlite::toJSON()}.}
}
\description{
Convert a \code{pxweb_query} object to a \code{json} string
}
