% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_as_rcode}
\alias{pxweb_query_as_rcode}
\title{Convert a \code{pxweb_query} object to R code}
\usage{
pxweb_query_as_rcode(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\description{
Convert a \code{pxweb_query} object to R code
}
\details{
One element per row is returned.
}
\keyword{internal}
