% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_as_v2}
\alias{pxweb_query_as_v2}
\title{Convert a \code{pxweb_query} object to a PXWEB API v2 query.}
\usage{
pxweb_query_as_v2(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\value{
a \code{pxweb_query_v2} object.
}
\description{
Convert a \code{pxweb_query} object to a PXWEB API v2 query.
}
\keyword{internal}
