% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{pxweb_query_dim_splittable}
\alias{pxweb_query_dim_splittable}
\title{Get vector indicating splittable variables}
\usage{
pxweb_query_dim_splittable(pxq, pxmd)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\value{
a named logical vector.
}
\description{
Get vector indicating splittable variables
}
\details{
Splitable variables are variables that can be split. Content variables cannot be split,
not variables with filter == "top"
}
\keyword{internal}
