% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_filter}
\alias{pxweb_query_filter}
\title{Get query values}
\usage{
pxweb_query_filter(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\value{
query variable selection filters as a named character vector.
}
\description{
Get query values
}
\keyword{internal}
