% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_list_as_v2}
\alias{pxweb_query_list_as_v2}
\title{Convert a named list to a PXWEB API v2 request.}
\usage{
pxweb_query_list_as_v2(x)
}
\arguments{
\item{x}{a named list with character values or \code{pxweb_query_selection}
helper objects.}
}
\value{
A list with a \code{body} slot containing a \code{pxweb_query_v2} object and
an \code{extra_query} slot containing v2 query parameters such as
\code{codelist[Variable]} and \code{outputValues[Variable]}.
}
\description{
Convert a named list to a PXWEB API v2 request.
}
\keyword{internal}
