% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_selection}
\alias{pxweb_query_selection}
\title{Construct a PXWEB API v2 query selection helper.}
\usage{
pxweb_query_selection(
  type,
  value_codes,
  codelist = NULL,
  output_values = NULL,
  class = character()
)
}
\arguments{
\item{type}{selection type.}

\item{value_codes}{values to use in the v2 \code{valueCodes} selection.}

\item{codelist}{optional codelist id.}

\item{output_values}{optional output values mode.}

\item{class}{additional S3 class names.}
}
\description{
Construct a PXWEB API v2 query selection helper.
}
\keyword{internal}
