% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_selection_extra_query}
\alias{pxweb_query_selection_extra_query}
\title{Build codelist-related query parameters for a v2 selection helper.}
\usage{
pxweb_query_selection_extra_query(x, variable_code)
}
\arguments{
\item{x}{a \code{pxweb_query_selection} object.}

\item{variable_code}{PXWEB variable code.}
}
\description{
Build codelist-related query parameters for a v2 selection helper.
}
\keyword{internal}
