% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_v2}
\alias{pxweb_query_v2}
\title{Construct a \code{pxweb_query_v2} object.}
\usage{
pxweb_query_v2(x)
}
\arguments{
\item{x}{a list with a PXWEB API v2 \code{selection} body.}
}
\value{
a \code{pxweb_query_v2} object.
}
\description{
Construct a \code{pxweb_query_v2} object.
}
\keyword{internal}
