% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_v2_request}
\alias{pxweb_query_v2_request}
\title{Construct a PXWEB API v2 request object.}
\usage{
pxweb_query_v2_request(body, extra_query = list())
}
\arguments{
\item{body}{a \code{pxweb_query_v2} object.}

\item{extra_query}{additional URL query parameters.}
}
\description{
Construct a PXWEB API v2 request object.
}
\keyword{internal}
