% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_value_as_v2}
\alias{pxweb_query_value_as_v2}
\title{Convert one query value to a PXWEB API v2 selection fragment.}
\usage{
pxweb_query_value_as_v2(x, variable_code)
}
\arguments{
\item{x}{a character vector or \code{pxweb_query_selection} object.}

\item{variable_code}{PXWEB variable code.}
}
\description{
Convert one query value to a PXWEB API v2 selection fragment.
}
\keyword{internal}
