% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_values}
\alias{pxweb_query_values}
\title{Get query filter}
\usage{
pxweb_query_values(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\value{
query variable selection values as a named list of character vectors.
}
\description{
Get query filter
}
\keyword{internal}
