% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_api_version}
\alias{pxweb_search_api_version}
\title{Detect the PXWEB API version used by a search URL}
\usage{
pxweb_search_api_version(api_url)
}
\arguments{
\item{api_url}{a PXWEB API URL.}
}
\description{
Detect the PXWEB API version used by a search URL
}
\keyword{internal}
