% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_empty_v1}
\alias{pxweb_search_empty_v1}
\title{Create an empty PXWEB API v1 search result}
\usage{
pxweb_search_empty_v1()
}
\description{
Create an empty PXWEB API v1 search result
}
\keyword{internal}
