% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_link}
\alias{pxweb_search_link}
\title{Extract a link URL from a PXWEB API v2 links list}
\usage{
pxweb_search_link(links, rel)
}
\arguments{
\item{links}{a PXWEB API links list.}

\item{rel}{the link relation to extract.}
}
\description{
Extract a link URL from a PXWEB API v2 links list
}
\keyword{internal}
