% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_num}
\alias{pxweb_search_num}
\title{Extract a scalar numeric value from a list}
\usage{
pxweb_search_num(x, name)
}
\arguments{
\item{x}{a list.}

\item{name}{the element name to extract.}
}
\description{
Extract a scalar numeric value from a list
}
\keyword{internal}
