% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_parse_v1}
\alias{pxweb_search_parse_v1}
\title{Parse PXWEB API v1 search results}
\usage{
pxweb_search_parse_v1(x, api_url)
}
\arguments{
\item{x}{a parsed PXWEB API v1 search response.}

\item{api_url}{the PXWEB API v1 database root URL used for the search.}
}
\description{
Parse PXWEB API v1 search results
}
\keyword{internal}
