% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_parse_v2}
\alias{pxweb_search_parse_v2}
\title{Parse PXWEB API v2 search results}
\usage{
pxweb_search_parse_v2(x)
}
\arguments{
\item{x}{a parsed PXWEB API v2 search response.}
}
\description{
Parse PXWEB API v2 search results
}
\keyword{internal}
