% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_path_parts}
\alias{pxweb_search_path_parts}
\title{Split a parsed URL path into parts}
\usage{
pxweb_search_path_parts(parsed_url)
}
\arguments{
\item{parsed_url}{a parsed URL object.}
}
\description{
Split a parsed URL path into parts
}
\keyword{internal}
