% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_request}
\alias{pxweb_search_request}
\title{Do a PXWEB search GET request}
\usage{
pxweb_search_request(url, query, ...)
}
\arguments{
\item{url}{a PXWEB search URL.}

\item{query}{a named list of query string parameters.}

\item{...}{further arguments passed to \code{httr::GET}.}
}
\description{
Do a PXWEB search GET request
}
\keyword{internal}
