% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_response}
\alias{pxweb_search_response}
\title{Parse a PXWEB search response}
\usage{
pxweb_search_response(response)
}
\arguments{
\item{response}{a \code{httr} response object.}
}
\description{
Parse a PXWEB search response
}
\keyword{internal}
