% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_trim_slash}
\alias{pxweb_search_trim_slash}
\title{Remove trailing slashes from a URL string}
\usage{
pxweb_search_trim_slash(x)
}
\arguments{
\item{x}{a character string.}
}
\description{
Remove trailing slashes from a URL string
}
\keyword{internal}
