% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_url}
\alias{pxweb_search_url}
\title{Build the PXWEB search URL}
\usage{
pxweb_search_url(api_url, version)
}
\arguments{
\item{api_url}{a PXWEB API root URL.}

\item{version}{PXWEB API version, \code{"v1"} or \code{"v2"}.}
}
\description{
Build the PXWEB search URL
}
\keyword{internal}
