% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_search.R
\name{pxweb_search_v1_table_url}
\alias{pxweb_search_v1_table_url}
\title{Build a PXWEB API v1 table URL from a search hit}
\usage{
pxweb_search_v1_table_url(api_url, path, id)
}
\arguments{
\item{api_url}{the PXWEB API v1 database root URL.}

\item{path}{the search hit path.}

\item{id}{the search hit table id.}
}
\description{
Build a PXWEB API v1 table URL from a search hit
}
\keyword{internal}
