% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{pxweb_split_query}
\alias{pxweb_split_query}
\title{Split query in optimal sub-queries}
\usage{
pxweb_split_query(pxq, px, pxmd)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{px}{a \code{pxweb} object.}

\item{pxweb_metadata}{a \code{pxweb_metadata} object.}
}
\value{
a list with \code{pxweb_query} objects.
}
\description{
Split query in optimal sub-queries
}
\details{
Computes (brute-force) the optimal split of a query to
match the api maximum value limit. It also take into account 
that time variables and content variables should not be split.
Also variables with filter "top" should not be split, since
the top filter does not supply the individual levels, just a 
number. This can probably be improved further.
}
\keyword{internal}
