% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_table_response_v2.R
\name{pxweb_table_response_v2}
\alias{pxweb_table_response_v2}
\title{Construct a \code{table_response_v2} object.}
\usage{
pxweb_table_response_v2(x)
}
\arguments{
\item{x}{a list returned from a PXWEB API to convert to a \code{table_response_v2} object.}
}
\value{
a \code{pxweb_table_v2} object.
}
\description{
An object that contain the data for a given PXWEB TableResponse (API v2).
}
\keyword{internal}
