% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb.R
\name{pxweb_tempdir}
\alias{pxweb_tempdir}
\title{Setup temorary directory for the pxweb}
\usage{
pxweb_tempdir(to = "apis")
}
\arguments{
\item{to}{to what part of the temp folder (apis or logs)}
}
\description{
Setup temorary directory for the pxweb
}
\keyword{internal}
