% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_test_api_endpoint.R
\name{pxweb_test_create_api_paths}
\alias{pxweb_test_create_api_paths}
\title{Create all paths from a list of pxweb_api_catalogue entries}
\usage{
pxweb_test_create_api_paths(apis)
}
\arguments{
\item{apis}{a list of pxweb_api_catalogue_entry elements}
}
\description{
Create all paths from a list of pxweb_api_catalogue entries
}
\keyword{internal}
