% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_test_api_endpoint.R
\name{pxweb_test_time_limit}
\alias{pxweb_test_time_limit}
\alias{is_test_time_limit_reached}
\title{Test time limit object}
\usage{
pxweb_test_time_limit(time_limit)

is_test_time_limit_reached(x)
}
\arguments{
\item{time_limit}{the number of seconds the API will be tested.}
}
\description{
Test time limit object
}
\details{
Used to limit testing of API:s.
}
\keyword{internal}
