% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_validate_query_with_metadata}
\alias{pxweb_validate_query_with_metadata}
\title{Validate a pxweb_query with pxweb_metadata}
\usage{
pxweb_validate_query_with_metadata(pxq, pxmd)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{pxmd}{a \code{pxweb_metadata} object.}
}
\description{
Validate a pxweb_query with pxweb_metadata
}
\details{
Assert that all query values exist in metadata.
}
\keyword{internal}
