% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pxweb_api_class.R
\docType{class}
\name{pxwebapi-class}
\alias{pxwebapi}
\alias{pxwebapi-class}
\title{A Reference Class to represent a pxweb api connection}
\description{
A Reference Class to represent a pxweb api connection
}
\section{Fields}{

\describe{
\item{\code{api_name}}{The name of the api (api domain)}

\item{\code{url}}{The URL to the api}

\item{\code{url_type}}{What type of pxweb api url ("database", "level", "table")}

\item{\code{api_limits}}{The limits of the api:
- max_values : maximum number of values to download in each call
- max_calls: maximum number of calls per time_window
- time_window: time window in seconds}

\item{\code{api_timer_file}}{The file path to api_timer.rdata file to store api_calls}

\item{\code{meta_data}}{Meta data of the file url}
}}
\section{Methods}{

\describe{
\item{\code{add_call_to_timer(calls = 1)}}{Add a call to the api to the api timer.}

\item{\code{fetch_api_config()}}{Get the api configuration from the pxweb api.}

\item{\code{fetch_metadata()}}{Fetch metadata from the pxweb api.}

\item{\code{get_metadata()}}{Get the metadata from the object.}

\item{\code{initialize(api_url = "")}}{Create a new pxwebapi object.}

\item{\code{set_api_timer_file()}}{Set file to use for the api timer.}

\item{\code{set_url_type()}}{Get the url type based on the meta data.}
}}
\examples{
  pxwebapi$new()
  pxwebapi$new(api_url = "http://api.scb.se/OV0104/v1/doris/sv/")
  pxwebapi$new(api_url = "http://api.scb.se/OV0104/v1/doris/sv/ssd/")
  pxwebapi$new(api_url = "http://api.scb.se/OV0104/v1/doris/sv/ssd/PR/PR0101/PR0101E/Basbeloppet")
  try(pxwebapi$new(api_url = "http://www.dn.se"))
  
}
\references{
http://www.scb.se/Grupp/OmSCB/API/API-description.pdf
}

