% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_get_pxweb_data_internal.R
\name{reorder_and_check_dims}
\alias{reorder_and_check_dims}
\title{Reorders the dimensions of a \code{get_pxweb_data()} call and checks that all dimensions are correct.}
\usage{
reorder_and_check_dims(url, dims)
}
\arguments{
\item{url}{Url to check against}

\item{dims}{dims object to check}
}
\value{
a correct (and checked) dims object
}
\description{
Reorders the dimensions of a \code{get_pxweb_data()} call and checks that all dimensions are correct.
}
\keyword{internal}
