% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_load_save.R
\name{save_pxweb}
\alias{save_pxweb}
\alias{load_pxweb}
\alias{load_pxweb_calls}
\alias{load_pxweb_config}
\alias{load_pxweb_api_subpath}
\title{Save and load \code{pxweb} objects from R temp folder}
\usage{
save_pxweb(obj)

load_pxweb(obj)

load_pxweb_calls(obj)

load_pxweb_config(obj)

load_pxweb_api_subpath(obj)
}
\arguments{
\item{obj}{a \code{pxweb} object}
}
\description{
Save and load \code{pxweb} objects from R temp folder
}
\keyword{internal}
