% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scbGetLevels}
\alias{scbGetLevels}
\title{Get levels from API node}
\usage{
scbGetLevels(baseURL, descriptions = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{baseURL}{The base URL to use, depending on the web service.}

\item{descriptions}{Whether to include node descriptions with the list of node IDs. (default: \code{FALSE})}

\item{quiet}{Quiet mode. Whether to stop with an error if the input node does not contain any subnodes. If set to \code{TRUE}, the function will quietly return FALSE without any errors. (default: \code{FALSE})}

\item{...}{further argument to send to \code{scbGetMetadata}}
}
\description{
Get levels from a node in the API. If at the lowest node, return a warning.
}
\examples{
lev <- scbGetLevels(baseURL = base_url("sweSCB", "v1", "sv"))
}

