% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_split_query.R
\name{split_dimensions_left_right}
\alias{split_dimensions_left_right}
\title{Split variables into chunks}
\usage{
split_dimensions_left_right(x, bool, max_size)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\value{
a \code{pxweb_split_dimensions}
}
\description{
Split variables into chunks
}
\details{
Splitable variables are variables that can be split. Content variables cannot be split,
not variables with filter == "top"
}
\keyword{internal}
