% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{str_pad}
\alias{str_pad}
\title{Pad a string to a fixed size}
\usage{
str_pad(txt, n = 5, pad = " ", type = "left")
}
\arguments{
\item{n}{final char width}

\item{pad}{pad symbol}

\item{type}{pad from 'left' or 'right'.}

\item{x}{a character vector to pad}
}
\description{
Pad a string to a fixed size
}
\keyword{internal}
