% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{str_trim}
\alias{str_trim}
\title{Taken from \code{trimws} for reasons of compatibility with previous R versios.}
\usage{
str_trim(x, which = c("both", "left", "right"))
}
\arguments{
\item{x}{a string to trim.}

\item{which}{how to trim the string.}
}
\description{
Taken from \code{trimws} for reasons of compatibility with previous R versios.
}
\seealso{
trimws
}
\keyword{internal}
