% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_api_catalogue.R
\name{temp_api_folder_path}
\alias{temp_api_folder_path}
\title{The path to the folder to store paths to api.json}
\usage{
temp_api_folder_path()
}
\description{
The path to the folder to store paths to api.json
}
\keyword{internal}
