% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_pxweb_api.R
\name{test_pxweb_api}
\alias{test_pxweb_api}
\title{Test a full or a part of an pxweb api.}
\usage{
test_pxweb_api(url, download_all = FALSE, seed = as.integer(Sys.time()))
}
\arguments{
\item{url}{The base url to the pxweb api (or a branch of the metadata tree)}

\item{download_all}{Should all data be downloaded (TRUE) or sample (FALSE)}

\item{seed}{Seed to use if download_all is FALSE and a sample is downloaded}
}
\value{
Function returns a data.frame with information on each node
Two variables are added:
\code{checked} : The node has been checked
\code{error} : Whether there were errors encountered with the call
\code{download_error} : Whether there were errors encountered during download
}
\description{
The function can be used to test a whole pxweb api by using the api base url.
By using a branch in a tree the api is tested below this branch.
}

