% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_test_pxweb_api_internal.R
\name{test_pxweb_api_get_data}
\alias{test_pxweb_api_get_data}
\title{Download data for each bottom node in the pxweb api}
\usage{
test_pxweb_api_get_data(nodes, nodesList, download_all = FALSE,
  seed = as.integer(Sys.time()))
}
\arguments{
\item{nodes}{Nodes downloaded with \link{test_pxweb_api_get_nodes}}

\item{nodesList}{Node metadata downloaded with \link{test_pxweb_api_get_node_metadata}}

\item{download_all}{Should all data be downloaded (TRUE) or sample (FALSE)}

\item{seed}{Seed to use if download_all is FALSE and a sample is downloaded}
}
\value{
data.frame with information \code{download_error} containing download errors
}
\description{
Download data for each bottom node in the pxweb api
}
\keyword{internal}
