% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_pxweb_api_get_node_metadata}
\alias{test_pxweb_api_get_node_metadata}
\title{Get metadata from all bottom nodes identified with \link{test_pxweb_api_get_nodes}.}
\usage{
test_pxweb_api_get_node_metadata(nodes)
}
\arguments{
\item{nodes}{A data.frame returned by \link{test_pxweb_api_get_nodes}}
}
\value{
List with one list per node object containing meta data
}
\description{
Get metadata from all bottom nodes identified with \link{test_pxweb_api_get_nodes}.
}

