% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_catalogue.R
\name{update_pxweb_apis}
\alias{update_pxweb_apis}
\title{Download and update the PXWEB API catalogue}
\usage{
update_pxweb_apis()
}
\description{
Download and update the PXWEB API catalogue
}
\details{
Downloads the latest api catalogue from github.
}
