% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_api_list}
\alias{write_api_list}
\title{Write api catalogue to json file}
\usage{
write_api_list(api_list)
}
\arguments{
\item{api_list}{api list in raw format to write to json file.}
}
\description{
Write api catalogue to json file
}

