% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct_functions.R
\docType{data}
\name{api_catalogue}
\alias{api_catalogue}
\alias{update_pxweb_apis}
\alias{api_parameters}
\alias{ApiData}
\alias{MakeUrl}
\alias{base_url}
\alias{get_pxweb_data}
\alias{get_pxweb_dims}
\alias{get_pxweb_levels}
\alias{get_pxweb_metadata}
\alias{pxweb_api}
\alias{checkForLevels}
\title{Defunct functions}
\format{
An object of class \code{list} of length 1.
}
\usage{
api_catalogue()

update_pxweb_apis()

api_parameters(url = NULL)

ApiData(
  urlToData,
  ...,
  getDataByGET = FALSE,
  returnMetaData = FALSE,
  returnMetaValues = FALSE,
  returnMetaFrames = FALSE,
  returnApiQuery = FALSE,
  defaultJSONquery = c(1, -2, -1),
  verbosePrint = FALSE,
  use_factors = FALSE,
  urlType = "SSB"
)

MakeUrl(id, urlType = "SSB", getDataByGET = FALSE)

base_url(api, version = NULL, language = NULL)

get_pxweb_data(url, dims, clean = FALSE, encoding = NULL)

get_pxweb_dims(node, verbose = TRUE)

get_pxweb_levels(baseURL, descriptions = FALSE, quiet = FALSE, ...)

get_pxweb_metadata(
  path = NULL,
  node = NULL,
  topnodes = NULL,
  quiet = TRUE,
  baseURL = NULL,
  ...
)

pxweb_api

checkForLevels(url)
}
\arguments{
\item{url}{Defunct argument.}

\item{urlToData}{Defunct argument.}

\item{...}{Defunct argument.}

\item{getDataByGET}{Defunct argument.}

\item{returnMetaData}{Defunct argument.}

\item{returnMetaValues}{Defunct argument.}

\item{returnMetaFrames}{Defunct argument.}

\item{returnApiQuery}{Defunct argument.}

\item{defaultJSONquery}{Defunct argument.}

\item{verbosePrint}{Defunct argument.}

\item{use_factors}{Defunct argument.}

\item{urlType}{Defunct argument.}

\item{id}{Defunct argument.}

\item{api}{Defunct argument.}

\item{version}{Defunct argument.}

\item{language}{Defunct argument.}

\item{dims}{Defunct argument.}

\item{clean}{Defunct argument.}

\item{encoding}{Defunct argument.}

\item{node}{Defunct argument.}

\item{verbose}{Defunct argument.}

\item{baseURL}{Defunct argument.}

\item{descriptions}{Defunct argument.}

\item{quiet}{Defunct argument.}

\item{path}{Defunct argument.}

\item{topnodes}{Defunct argument.}
}
\description{
These function has as from version 0.10.0 become defunct.
Call the functions to get information on new functions to use.
}
\keyword{datasets}
