% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_catalogue.R
\name{api_catalogue}
\alias{api_catalogue}
\title{Get the api catalogue}
\usage{
api_catalogue()
}
\description{
Get the api catalogue
}
\details{
Get the catalogue of api and stor it as a list.
}
\examples{
api_catalogue()
}

