% Generated by roxygen2 (4.0.1): do not edit by hand
\name{api_parameters}
\alias{api_parameters}
\title{Return options for database, version and language choices}
\usage{
api_parameters(...)
}
\arguments{
\item{...}{Additional parameters; currently ignored}
}
\description{
...
}
\examples{
api_parameters()
}

