% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_parameters.R
\name{api_parameters}
\alias{api_parameters}
\title{Return options for database, version and language choices}
\usage{
api_parameters(url = NULL)
}
\arguments{
\item{url}{Get parameters for the api with this url (NULL = get all API configs)}
}
\description{
...
}
\examples{
api_parameters()
}
\keyword{internal}
