% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_table_response_v2.R
\name{assert_pxweb_table_response_v2}
\alias{assert_pxweb_table_response_v2}
\title{Assert that x is a correct \code{pxweb_table_response_v2} object.}
\usage{
assert_pxweb_table_response_v2(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Assert that x is a correct \code{pxweb_table_response_v2} object.
}
\keyword{internal}
