% Generated by roxygen2 (4.0.1): do not edit by hand
\name{base_url}
\alias{base_url}
\title{Return base URL to API}
\usage{
base_url(api, version = "v1", lang = "en", ...)
}
\arguments{
\item{api}{API provider ('api.scb.se' or 'pxwebapi2.stat.fi')}

\item{version}{The version of PXWEB API to use. (Default: \code{v1})}

\item{lang}{The language (two letters) to use in the fetched data. (Default: \code{en})}

\item{...}{Additional parameters. These are currently ignored.}
}
\description{
...
}
\examples{
a <- base_url("api.scb.se", "v1", "sv")
print(a)
}

