% Generated by roxygen2 (4.0.1): do not edit by hand
\name{base_url}
\alias{base_url}
\title{Return base URL to API}
\usage{
base_url(database, version = "v1", lang = "en", ...)
}
\arguments{
\item{database}{Source database ('sweSCB' or 'statfi')}

\item{version}{The version of SCB API to use. (Default: \code{v1})}

\item{lang}{The language (two letters) to use in the fetched data. (Default: \code{sv})}

\item{...}{Additional parameters. These are currently ignored.}
}
\description{
...
}
\examples{
a <- base_url("sweSCB", "v2", "en")
print(a)
}

