% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_url.R
\name{base_url}
\alias{base_url}
\title{Return base URL to API}
\usage{
base_url(api, version = NULL, language = NULL)
}
\arguments{
\item{api}{API provider ('api.scb.se' or 'pxwebapi2.stat.fi')}

\item{version}{The version of PXWEB API to use.}

\item{language}{The language (two letters) to use in the fetched data.}

\item{...}{Additional parameters. These are currently ignored.}
}
\description{
...
}
\examples{
a <- base_url("api.scb.se")
print(a)

}
\keyword{internal}

