% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_api_catalogue.R
\name{check_new_pxweb_apis}
\alias{check_new_pxweb_apis}
\title{Check and compare differences between remote and local api catalogue}
\usage{
check_new_pxweb_apis()
}
\description{
Check and compare differences between remote and local api catalogue
}
\keyword{internal}
