% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_pxweb}
\alias{clean_pxweb}
\title{Clean raw data from PX-WEB}
\usage{
clean_pxweb(data2clean, url, dims, content_node = NULL)
}
\arguments{
\item{data2clean}{Data to clean.}

\item{url}{url to the bottom nod (to get meta data)}

\item{content_node}{node with content downloaded with \link{get_pxweb_metadata}.
If NULL, meta data is downloaded with \link{get_pxweb_metadata}.}
}
\value{
data frame melted and in R (numeric) format
}
\description{
This function clean the raw data from PX-WEB to R tall format.
}

