% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{create_batch_list}
\alias{create_batch_list}
\title{The function takes an url and a dims object and calculates if this needs to
be downloaded in batches due to download limit restrictions in the pxweb api.
If '*' is used it will get the numberof values for this dimension using a 
get_pxweb_metadata call.}
\usage{
create_batch_list(url, dims)
}
\arguments{
\item{url}{The url to download from.}

\item{dims}{The dimension object to use for downloading}
}
\description{
The function takes an url and a dims object and calculates if this needs to
be downloaded in batches due to download limit restrictions in the pxweb api.
If '*' is used it will get the numberof values for this dimension using a 
get_pxweb_metadata call.
}
\keyword{internal}

