% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deparseLevels}
\alias{deparseLevels}
\title{Functions from sweSCB package that are not called anywhere (yet)
but might be useful later
Function to deparse an URL into its components.}
\usage{
deparseLevels(place, returnDistance = FALSE, baseURL, ...)
}
\arguments{
\item{place}{Location in hierarchy, in the form of a (data-node complete) URL.}

\item{returnDistance}{Whether to return only the distance (in nodes) from top node to URL. (Default: \code{FALSE})}

\item{baseURL}{The base URL to use, depending on the web service.}

\item{...}{Further arguments passed to \code{base_url()}.}
}
\description{
This function is currently not used in the package.
}

