% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_api_index}
\alias{get_api_index}
\title{Get index of which api has the name or teh alias}
\usage{
get_api_index(api_name, api_list)
}
\arguments{
\item{api_name}{pxweb api name or alias to lookup.}

\item{api_list}{pxweb api list created with \code{get_api_list()}}
}
\description{
Get index of which api has the name or teh alias
}

