% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_catalogue.R
\name{get_api_list_remote}
\alias{get_api_list_remote}
\title{Get the api catalogue from github}
\usage{
get_api_list_remote(raw = FALSE)
}
\arguments{
\item{raw}{Get the raw list from the api json file.}
}
\value{
api_list object
}
\description{
Get the api catalogue from github
}
\keyword{internal}
