% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_pxweb_dims}
\alias{get_pxweb_dims}
\title{Get data dimensions form a bottom node in SCB API}
\usage{
get_pxweb_dims(node, verbose = TRUE)
}
\arguments{
\item{node}{Bottom node to deparse into object}

\item{verbose}{Verbose output mode.}
}
\description{
Deparse and reorder data form the metadata object for pretty output
}
\examples{
## CONTINUED FROM EXAMPLES IN get_pxweb_metadata()
# Get metadata for a named variable
baseURL <- base_url("sweSCB", "v1", "sv")
url <- paste(c(baseURL,"AM","AM0102","AM0102A","KLStabell14LpMan"), collapse="/")
metadata <- get_pxweb_metadata(url)

# Get dimensions (names of dimensions are printed in the terminal)
dims <- get_pxweb_dims(metadata)

# Get data
test <- get_pxweb(metadata$URL, dims=list(
   Myndighet = "C02",
   Kon = "*",
   Heltiddeltid = "*",
   ContentsCode = "*",
   Tid = "*"
))
}
\seealso{
\link{get_pxweb_metadata}
}

