% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_pxweb_dims.R
\name{get_pxweb_dims}
\alias{get_pxweb_dims}
\title{Get data dimensions form a bottom node in a PX-Web/PC-Axis API}
\usage{
get_pxweb_dims(node, verbose = TRUE)
}
\arguments{
\item{node}{Bottom node to deparse into object}

\item{verbose}{Verbose output mode.}
}
\description{
Deparse and reorder data form the metadata object for pretty output
}
\examples{
\dontrun{
# Download meta data
bottom_node <- get_pxweb_metadata("http://api.scb.se/OV0104/v1/doris/sv/ssd/AM/AM0114/LCIArbKv")

# Get dimensions (names of dimensions are printed in the terminal)
dims <- get_pxweb_dims(bottom_node)
}
}
\seealso{
\link{get_pxweb_metadata}
}

