% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_add_config.R
\name{http_was_redirected}
\alias{http_was_redirected}
\title{http_was_redirected}
\usage{
http_was_redirected(r)
}
\arguments{
\item{response}{an httr response object, e.g. from a call to httr::GET()}
}
\value{
list with slots \code{was_redirected}, \code{redirected_from} and  \code{redirected_to}
}
\description{
http_was_redirected
}
\examples{
\dontrun{
  r <- httr::GET("http://httpbin.org/redirect/2")
  pxweb:::http_was_redirected(r)
}
}
\references{
Function in large parts taken from \url{https://petermeissner.de/blog/2018/11/07/using-httr-to-detect-redirects/}.
}
\keyword{internal}
