% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interactive_pxweb}
\alias{interactive_pxweb}
\title{Find and download data interactively from PX-WEB API}
\usage{
interactive_pxweb(api, version = NULL, language = NULL, history = FALSE,
  ...)
}
\arguments{
\item{api}{The name of the pxweb api to connect to or an \link{pxweb_api} object.}

\item{version}{The version of the pxweb api to use. Default is pxweb api default, see \link{api_parameters}.}

\item{language}{The language of the pxweb api to use. Default is english.}

\item{history}{keep the history when the function is running.}

\item{...}{further parameters. These are currently ignored.}
}
\description{
Wrapper function (for \link{get_pxweb_data} and \link{get_pxweb_metadata}) to simply find and download data to the current R session.
}
\examples{
api_parameters() # List apis
\donttest{
 d <- interactive_pxweb("scb")
 d <- interactive_pxweb(api = "statistik.sjv.se")
 d <- interactive_pxweb("scb", language = "sv")
}
}
\seealso{
\code{\link{get_pxweb_metadata}}, \code{\link{get_pxweb_data}}
}

