% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data.R
\name{pxd_values_to_valuetexts}
\alias{pxd_values_to_valuetexts}
\title{Convert a pxweb data objects values to valuetext}
\usage{
pxd_values_to_valuetexts(x, variable_code, variable_vector)
}
\arguments{
\item{x}{a \code{pxweb_data} object}

\item{variable_code}{the variable name of the variable to convert. NOTE! Need to be the variable code.}

\item{variable_vector}{the vector with the variables to convert.}
}
\description{
Convert a pxweb data objects values to valuetext
}
\keyword{internal}
