% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_allowed_input}
\alias{pxe_allowed_input}
\alias{pxe_allowed_input_df}
\alias{pxe_allowed_input.pxweb_explorer}
\title{Defines allowed input for a position in a \code{pxweb_explorer} or character object.}
\usage{
pxe_allowed_input(x)

pxe_allowed_input_df()

\method{pxe_allowed_input}{pxweb_explorer}(x)
}
\arguments{
\item{x}{a object to get allowed input for.}
}
\description{
Defines allowed input for a position in a \code{pxweb_explorer} or character object.
}
\keyword{internal}
