% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxe_input}
\alias{pxe_input}
\title{Get (allowed) inputs for a \code{pxweb_explorer} object.}
\usage{
pxe_input(pxe)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer_object}}
}
\description{
Get (allowed) inputs for a \code{pxweb_explorer} object.
}
\keyword{internal}
