% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb-package.R
\docType{package}
\name{pxweb-package}
\alias{pxweb-package}
\title{Interface to PX-WEB APIs from R}
\description{
PXWEB is a common web API used by many European Statistical agencies to disseminate official statistics.
The pxweb package facilitates connections and usage of these APIs.
}
\references{
The decription of the PXWEB API here:
\url{https://www.scb.se/en/About-us/about-the-website-and-terms-of-use/open-data-api/api-for-the-statistical-database/}
The official home page of PXWEB can be found here:
\url{https://www.scb.se/sv_/PC-Axis/Programs/PX-Web/}
}
