% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb-package.R
\docType{package}
\name{pxweb-package}
\alias{pxweb-package}
\title{pxweb}
\description{
Interface to the PX-WEB API
}
\references{
https://www.scb.se/en/About-us/about-the-website-and-terms-of-use/open-data-api/api-for-the-statistical-database/
Documentation
PXWEB home page
}
