% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_add_metadata_to_query}
\alias{pxweb_add_metadata_to_query}
\alias{pxweb_remove_metadata_from_query}
\title{Add and remove metadata to query}
\usage{
pxweb_add_metadata_to_query(pxq, pxmd)

pxweb_remove_metadata_from_query(pxq, pxmd)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{pxmd}{a \code{pxweb_metadata} object.}
}
\description{
Add and remove metadata to query
}
\details{
Add metadata values to variables with a query with filter "all".
}
\keyword{internal}
