% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_as_data_frame.R
\name{pxweb_as_data_frame}
\alias{pxweb_as_data_frame}
\alias{pxweb_as_data_frame.pxweb_data}
\alias{pxweb_as_data_frame.pxweb_data_comments}
\alias{pxweb_as_data_frame.pxweb_data_comment}
\alias{pxweb_as_data_frame.pxweb_levels}
\alias{as.data.frame.pxweb_data}
\alias{as.data.frame.pxweb_levels}
\title{Coerce to a \code{data.frame}}
\usage{
pxweb_as_data_frame(x, row.names = NULL, optional = FALSE, ...,
  stringsAsFactors = default.stringsAsFactors(),
  column.name.type = "text", variable.value.type = "text")

\method{pxweb_as_data_frame}{pxweb_data}(x, row.names = NULL,
  optional = FALSE, ..., stringsAsFactors = default.stringsAsFactors(),
  column.name.type = "text", variable.value.type = "text")

\method{pxweb_as_data_frame}{pxweb_data_comments}(x, optional = FALSE,
  ..., stringsAsFactors = default.stringsAsFactors())

\method{pxweb_as_data_frame}{pxweb_data_comment}(x, optional = FALSE,
  ..., stringsAsFactors = default.stringsAsFactors())

\method{pxweb_as_data_frame}{pxweb_levels}(x, optional = FALSE, ...,
  stringsAsFactors = default.stringsAsFactors())

\method{as.data.frame}{pxweb_data}(x, row.names = NULL,
  optional = FALSE, ..., stringsAsFactors = default.stringsAsFactors(),
  column.name.type = "text")

\method{as.data.frame}{pxweb_levels}(x, row.names = NULL,
  optional = FALSE, ..., stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{x}{an object to convert to \code{data.frame}.}

\item{row.names}{See \code{\link[base]{as.data.frame}}.}

\item{optional}{See \code{\link[base]{as.data.frame}}.}

\item{...}{See \code{\link[base]{as.data.frame}}.}

\item{stringsAsFactors}{See \code{\link[base]{as.data.frame}}.}

\item{column.name.type}{character: should \code{code} or \code{text} be used as column names?}

\item{variable.value.type}{character: should \code{code} or \code{text} be used as values in columns?}
}
\description{
Coerce to a \code{data.frame}
}
\seealso{
\code{\link[base]{as.data.frame}}.
}
\keyword{internal}
