% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_as_data_frame.R
\name{pxweb_as_data_frame}
\alias{pxweb_as_data_frame}
\alias{pxweb_as_data_frame.pxweb_data}
\title{Convert object to a \code{data.frame}}
\usage{
pxweb_as_data_frame(x, column_names = "text")

\method{pxweb_as_data_frame}{pxweb_data}(x, column_names = "text")
}
\arguments{
\item{x}{an object to convert.}

\item{column_names}{should \code{code} or \code{text} be used as column names?}
}
\description{
Convert object to a \code{data.frame}
}
\keyword{internal}
