% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_as_json.R
\name{pxweb_as_json}
\alias{pxweb_as_json}
\alias{pxweb_as_json.pxweb_query}
\title{Convert object to json}
\usage{
pxweb_as_json(x)

\method{pxweb_as_json}{pxweb_query}(x)
}
\arguments{
\item{x}{an object to convert.}
}
\description{
Convert object to json
}
\keyword{internal}
