% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_cite.R
\name{pxweb_cite}
\alias{pxweb_cite}
\title{Cite a PXWEB data object}
\usage{
pxweb_cite(x, style = "citation")
}
\arguments{
\item{x}{a \code{pxweb_data} object to cite.}

\item{style}{see \code{\link[utils]{bibentry}}.}
}
\description{
Cite a PXWEB data object
}
\details{
Functionality to automatic cite PXWEB data objects.
}
