% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data.R
\name{pxweb_data_colnames}
\alias{pxweb_data_colnames}
\title{Get query filter}
\usage{
pxweb_data_colnames(pxd, type = "text")
}
\arguments{
\item{pxd}{a \code{pxweb_data} object.}
}
\value{
character vector with column names.
}
\description{
Get query filter
}
\keyword{internal}
