% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_comments.R
\name{pxweb_data_column_comment}
\alias{pxweb_data_column_comment}
\alias{pxweb_data_value_comment}
\alias{pxweb_data_obs_comment}
\title{Construct a \code{pxweb_data_comment} object}
\usage{
pxweb_data_column_comment(x)

pxweb_data_value_comment(x, column_code, column_text)

pxweb_data_obs_comment(x, column_code, column_text)
}
\arguments{
\item{x}{an object to convert to a \code{pxweb_data_comment} object}
}
\description{
Construct a \code{pxweb_data_comment} object
}
\keyword{internal}
