% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_data_comments.R
\name{pxweb_data_comments}
\alias{pxweb_data_comments}
\alias{pxweb_data_comments.pxweb_data}
\title{Construct a \code{pxweb_data_comments} object.}
\usage{
pxweb_data_comments(x)

\method{pxweb_data_comments}{pxweb_data}(x)
}
\arguments{
\item{x}{a \code{pxweb_data} object.}
}
\value{
a \code{pxweb_data_comments} object
}
\description{
An object that contain the comments for a given PXWEB table.
}
