% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_interactive.R
\name{pxweb_interactive_input}
\alias{pxweb_interactive_input}
\title{Get input from user}
\usage{
pxweb_interactive_input(pxe, test_input = NULL)
}
\arguments{
\item{pxe}{a \code{pxweb_explorer} object to get user input for.}

\item{test_input}{supplying a test input (for testing only).}
}
\description{
Get input from user
}
\keyword{internal}
