% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_parse_response.R
\name{pxweb_parse_response}
\alias{pxweb_parse_response}
\alias{is_pxweb_response}
\title{Parse the response from a PXWEB API (advanced)}
\usage{
pxweb_parse_response(x)

is_pxweb_response(x)
}
\arguments{
\item{x}{a \code{httr} response object from a PXWEB call.}
}
\description{
The function parses the response from a call made to a PXWEB API
using the \code{httr} R package. In this way it is possible to parse the 
content of calls made outside the pxweb R package.
}
