% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_as_rcode}
\alias{pxweb_query_as_rcode}
\title{Print a \code{pxweb_query} object as R code}
\usage{
pxweb_query_as_rcode(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\description{
Print a \code{pxweb_query} object as R code
}
\seealso{
\code{\link{pxweb_query_as_json}}, \code{\link{pxweb_query}}
}
