% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query.R
\name{pxweb_query_dim}
\alias{pxweb_query_dim}
\title{Compue the dimension of the query}
\usage{
pxweb_query_dim(pxq)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}
}
\description{
Compue the dimension of the query
}
\keyword{internal}
