% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_helpers}
\alias{pxweb_query_helpers}
\alias{pxweb_all}
\alias{pxweb_latest}
\alias{pxweb_aggregation}
\alias{pxweb_valueset}
\alias{pxweb_top}
\alias{pxweb_bottom}
\title{PXWEB API v2 query helpers}
\usage{
pxweb_all()

pxweb_latest(code = "9999")

pxweb_aggregation(
  id,
  value_codes = "*",
  output_values = c("aggregated", "single")
)

pxweb_valueset(
  id,
  value_codes = "*",
  output_values = c("aggregated", "single")
)

pxweb_top(n)

pxweb_bottom(n)
}
\arguments{
\item{code}{value code used to represent the latest time period before it is
resolved against table metadata.}

\item{id}{codelist id, for example an aggregation id returned by a PXWEB API
v2 metadata response.}

\item{value_codes}{value codes to request from the selected codelist.}

\item{output_values}{whether the API should return aggregated values or
single values for the codelist.}

\item{n}{number of values to request.}
}
\value{
A \code{pxweb_query_selection} object.
}
\description{
Helper functions for expressing PXWEB API v2 query selections without
relying on magic strings in user code.
}
\examples{
query <- list(
  Region = pxweb_all(),
  Alder = pxweb_aggregation("agg_Alder5ar_1"),
  Tid = pxweb_latest()
)

}
