% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_v2_has_codelist}
\alias{pxweb_query_v2_has_codelist}
\title{Test whether a query variable uses a PXWEB API v2 codelist.}
\usage{
pxweb_query_v2_has_codelist(pxq, variable_code)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{variable_code}{PXWEB variable code.}
}
\description{
Test whether a query variable uses a PXWEB API v2 codelist.
}
\keyword{internal}
