% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_query_v2_helpers.R
\name{pxweb_query_v2_selection_type}
\alias{pxweb_query_v2_selection_type}
\title{Get the PXWEB API v2 selection type for a query variable.}
\usage{
pxweb_query_v2_selection_type(pxq, variable_code)
}
\arguments{
\item{pxq}{a \code{pxweb_query} object.}

\item{variable_code}{PXWEB variable code.}
}
\description{
Get the PXWEB API v2 selection type for a query variable.
}
\keyword{internal}
