% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_pxweb_api_get_nodes}
\alias{test_pxweb_api_get_nodes}
\title{Get the whole node tree from a given pxweb api}
\usage{
test_pxweb_api_get_nodes(url)
}
\arguments{
\item{url}{The base url to the pxweb api}
}
\value{
Function returns a data.frame with information on each node
Two variables are added:
\code{checked} : The node has been checked
\code{error} : Whether there were errors encountered with the call
}
\description{
Get the whole node tree from a given pxweb api
}

