% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update_pxweb_apis}
\alias{update_pxweb_apis}
\title{Download and update the PXWEB API catalogue}
\usage{
update_pxweb_apis()
}
\description{
Download and update the PXWEB API catalogue
}
\details{
Downloads the latest api catalogue from github.
}

